pyproject.toml
[tool.poetry]
name = "de-sprint-9"
version = "0.1.0"
description = ""
authors = ["Grishenkov Leonid <grishenkovleonid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg = {extras = ["binary"], version = "^3.1.10"}

[tool.poetry.group.dev.dependencies]
pgcli = "^3.5.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
ipykernel = "^6.25.1"
pandas = "^2.1.0"
tabulate = "^0.9.0"
s3fs = "2023.6.0"

[tool.poetry.group.apps-common.dependencies]
pandas = "2.1.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
kafka-python = "^2.0.2"
redis = "^4.6.0"

[tool.poetry.group.producer-app.dependencies]
s3fs = "2023.6.0"
pyarrow = "12.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
